requirements.txt
asgiref==3.8.1
black==25.1.0
click==8.2.1
Django==5.2.1
django-filter==24.1
djangorestframework==3.15.2
drf-yasg==1.21.7
Faker==24.4.0
importlib-metadata==6.8.0
inflection==0.5.1
iniconfig==2.1.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
psycopg2-binary==2.9.10
Pygments==2.19.1
pytest==8.4.0
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
uritemplate==4.1.1
zipp==3.22.0
